Create mocks file for unit tests
Move object multiplex type declaration.
