+++ app/scripts/desktop/storage.test.ts
@@ -2,7 +2,7 @@ import { randomBytes } from 'crypto';
 import Store from 'electron-store';
 import keytar from 'keytar';
 import ObfuscatedStore from './storage';
-import { DATA_2_MOCK, PASSWORD_MOCK, STORE_MOCK } from './test/utils';
+import { DATA_2_MOCK, PASSWORD_MOCK, STORE_MOCK } from './test/mocks';
 
 jest.mock('electron-store', () => jest.fn(), { virtual: true });
 jest.mock('crypto', () => ({ randomBytes: jest.fn() }), { virtual: true });
